+++ compiler/src/closure/ast.rs
@@ -84,7 +84,7 @@ pub struct Atom {
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub enum AtomKind {
     Reference(Reference),
-    Integer(i32),
+    Integer(i128),
     Unit,
 }
